Revert "Revert "APIMF-1777: Refactors and recursion control in RdfModelParser""
