+++ shared/src/main/scala/amf/core/rdf/graph/NodeFinder.scala
@@ -1,0 +1,12 @@
+package amf.core.rdf.graph
+
+import amf.core.rdf.{PropertyObject, RdfModel, Uri}
+
+class NodeFinder(graph: RdfModel) {
+  def findLink(property: PropertyObject) = {
+    property match {
+      case Uri(v) => graph.findNode(v)
+      case _      => None
+    }
+  }
+}
