+++ shared/src/main/scala/amf/core/rdf/parsers/DynamicLiteralParser.scala
@@ -1,0 +1,13 @@
+package amf.core.rdf.parsers
+
+import amf.core.model.domain.ScalarNode
+import amf.core.rdf.Literal
+
+object DynamicLiteralParser {
+  def parse(l: Literal): ScalarNode = {
+    val result = ScalarNode()
+    l.literalType.foreach(t => result.withDataType(t))
+    result.withValue(l.value)
+    result
+  }
+}
